There is now support for much many more operations
You are able to add values and even place in mathematical equations

THough I have to admit I found it very difficult finding a balance
for family wealth with this cubic function and it just ended up being
way to difficult to scale so I ended up just putting in plus family
wealth instead.

ANyways, there is defnetley a fremwork that could be used to make
config files at this point, I'm going to try to add conditionals and
finfish the default config before I merge this into master.