Cleaned everything up a little
Everything will run, but right now dnd-gen does not call
any script except namer, and it only calls namer to get a human
name.  We just need to add that functionality but it should not
be much of a pain.  It looks like it will be much easier to develop
for this from now on (assuming we can keep it clean)